Ssome fixes to the hook for embedded DB upgrade.
