+++ resources/bin/upgrade/hooks/PostUpgradeHooks.groovy
@@ -27,7 +27,7 @@ class PostUpgradeHooks {
 
     private static final Map<String, List<PostUpgradeHook>> ALL_HOOKS = [
             'authoring 3.0.x': [
-                    new UpgradeDbHook(),
+                    new UpgradeEmbeddedDbHook(),
                     new UpdateIndexIdFormatInPreviewTargetsHook(),
                     new EnableCrafterSearchInTargetsHook(),
                     new StartCrafterHook(['withSolr']),
@@ -35,13 +35,13 @@ class PostUpgradeHooks {
                     new RecreateSolrCoresHook()
             ],
             'delivery 3.0.x': [
-                    new UpgradeDbHook(),
+                    new UpgradeEmbeddedDbHook(),
                     new EnableCrafterSearchInTargetsHook(),
                     new StartCrafterHook(['withSolr']),
                     new RecreateSolrCoresHook()
             ],
             'authoring 3.1.0': [
-                    new UpgradeDbHook()
+                    new UpgradeEmbeddedDbHook()
             ]
     ]
